+++ packages/youtube_player_iframe/lib/src/youtube_player_scaffold.dart
@@ -6,9 +6,12 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:youtube_player_iframe/youtube_player_iframe.dart';
 
+import 'controller/youtube_player_controller.dart';
+import 'helpers/youtube_value_builder.dart';
+import 'helpers/youtube_value_provider.dart';
 import 'player_value.dart';
+import 'youtube_player.dart';
 
 /// A widget the scaffolds the [YoutubePlayer] so that it can be moved around easily in the view
 /// and handles the fullscreen functionality.
@@ -32,7 +35,8 @@ class YoutubePlayerScaffold extends StatefulWidget {
     ],
     this.enableFullScreenOnVerticalDrag = true,
     this.backgroundColor,
-    this.userAgent,
+    @Deprecated('Unused parameter. Use `YoutubePlayerParam.userAgent` instead.')
+        this.userAgent,
   });
 
   /// Builds the child widget.
@@ -100,8 +104,8 @@ class _YoutubePlayerScaffoldState extends State<YoutubePlayerScaffold> {
         controller: widget.controller,
         aspectRatio: widget.aspectRatio,
         gestureRecognizers: widget.gestureRecognizers,
-        backgroundColor: widget.backgroundColor,
         enableFullScreenOnVerticalDrag: widget.enableFullScreenOnVerticalDrag,
+        backgroundColor: widget.backgroundColor,
       ),
     );